Migrate logic to helper, move to using copyto instead of copying text and add unit tests
